core-465 Refactor code.
