+++ react-components/src/analysis.js
@@ -1,4 +1,4 @@
 import AnalysesView from "./analysis/view/AnalysesView";
-import ViceLogsViewer from "./analysis/view/ViceLogsViewer";
+import ViceLogsViewer from "./analysis/view/dialogs/ViceLogsViewer";
 
 export { AnalysesView, ViceLogsViewer };
